Move geocoder http client to http namespace
