+++ lib/geocoder.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module Geocoder
+  class Error < StandardError; end
+end
