+++ lib/geocoder/http/api.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Geocoder
+  module HTTP
+    class API
+      include Import[client: "geocoder.http.client"]
+
+      def coordinates(city)
+        client.request(:post, "coordinates", { city: city })
+      end
+    end
+  end
+end
